drm/amd/dal: refactor secondary pipe acquire logic
also fix tree_cfg index used in add/remove dpp in
mpc.c

Acked-by: Jordan Lazare <Jordan.Lazare@amd.com>

+++ drivers/gpu/drm/amd/dal/dc/dce100/dce100_resource.c
@@ -651,7 +651,7 @@ enum dc_status dce100_validate_bandwidth(
 enum dc_status dce100_validate_with_context(
 		const struct core_dc *dc,
 		const struct dc_validation_set set[],
-		uint8_t set_count,
+		int set_count,
 		struct validate_context *context)
 {
 	struct dc_context *dc_ctx = dc->ctx;
