+++ drivers/gpu/drm/amd/dal/dc/dce80/dce80_resource.c
@@ -606,7 +606,7 @@ enum dc_status dce80_validate_bandwidth(
 enum dc_status dce80_validate_with_context(
 		const struct core_dc *dc,
 		const struct dc_validation_set set[],
-		uint8_t set_count,
+		int set_count,
 		struct validate_context *context)
 {
 	struct dc_context *dc_ctx = dc->ctx;
